ideahub_app/pyproject.toml
[project]
name = "ideahub"
version = "0.1.0"
description = "IdeaHub - A material card management application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "Pillow>=10.0.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "websockets>=12.0",
    "aiohttp>=3.13.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
